refactor: unify component imports
